Implemented identity system and data configuration enhancements
In this commit, an identity system was implemented by adding a new 'ApplicationUser' class and modifying the project files to include essential packages for identity functionality. Also, the application's data configuration was enhanced by adding connection strings in the development appsettings file and establishing database context in new 'ApplicationDbContext' and 'Extensions' classes.